move to fluke sub-folder
