+++ src/main/java/com/ecostats/flukes/TraceCompare.java
@@ -17,7 +17,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
  */
-package com.ecostats;
+package com.ecostats.flukes;
 
 
 import java.io.Serializable;
