+++ src/main/java/com/ecostats/flukes/package-info.java
@@ -17,4 +17,4 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
  */
-package com.ecostats;
+package com.ecostats.flukes;
